Create The Big One
Remember that time when we've tried identifying the largest digit among the given integer? Well, let's take it to the next level, and figure out the same thing on arrays!



Let's do this one more time!



Instructions:

Input a non-zero positive integer. Then create an integer array that has a size equal to the inputted integer.
Using a loop, add integer values into the array one by one. The number of values to be stored is dependent on the inputted positive integer given in the first instruction.
Print out the elements of the array in reverse order.
Finally, print out the largest integer among all the elements on the array.
Input

1. Size of the array

2. Elements of the array

Output

The first line will contain a message prompt to input the size of the array.
The second line will prompt for the elements in the array.
The third line will contain the elements of the array in reversed order.
The last line contains the largest element.